requirements.txt
async-timeout==4.0.3
click==8.1.7
hiredis==2.3.2
pillow==10.3.0
python-dotenv==1.0.1
redis==5.0.4
